DFE-704 Decouple ConfirmContext from wizard step components and ModalConfirm
This commit removes the dependency on `ConfirmContext` in various
components (`ModalConfirm`, `WizardStepFormActions`, `WizardStepHeading`)
by making the `ConfirmContext.askConfirm` method async.

This allows us to hoist its invocation up to `TableToolPage` and work
with it in the new `Wizard.onStepChange` handler.

The benefit of this is that we can more easily test our components
without `ConfirmContext` and can refactor our code so that we don't need
to invoke `askConfirm` in multiple places.