PP-117 Add `jsRequired` flag to Details
This optimizes Details so that it can automatically render with all of
its progressive enhancements when `jsRequired` is true.

We require this optimization as it avoids IE11/Edge from momentarily
rendering the non-JS version seen in PP-117.

We can assume that it is currently safe to use in table tool, although
this may need re-work if we ever create a non-JS version.