refactor: MultiHeaderTable with improved header props
This commit significantly refactors and improves on the current
MultiHeaderTable implementation. New header prop types are introduced
so that we can more nicely describe how the headers should render.

Crucially we improve the handling of header 'subgroups' so it is more
obvious how we treat them differently to regular header groups.