EES-1424 Apply `mapTableHeadersConfig` to all relevant call sites
This commit fixes table header configuration being passed to
`TimePeriodDataTable` in their 'raw' form (e.g. directly from an API
response). This causes table rows to be populated with n/a as the
headers do not have their type information attached.

This is a regression due to earlier refactoring which removed
`mapTableHeadersConfig` from being applied in `TimePeriodDataTable`
itself.