EES-1458 Upgrade Next@9.2.3.canary.22 and core-js@3.6.4
This commit upgrades Next to a canary version so that we can gain access
to their new polyfill functionality. This actually removes the need for
core-js in the frontend, but we keep it so that we can transpile code
for our tests correctly.

We also upgrade core-js to v3 across the project so that we can remove
the nasty warnings about not specifying `corejs: 2` in our Babel configs