EES-778 Re-work ChartDataSelector to use multiple filter selects
This commit replaces the current form implementation with one that
allows the user to more granularly select their filter options.
Previously, the user would get one big 'Filters' select that would
list all of the filter combinations as a comma separated string which is
not very optimal in terms of user experience.

Additionally, we have refactored the form to also use Formik, which
adds validation and error handling that was previously missing.