EES-1073 Lift rendering of Legend to ChartRenderer
This commit implements a hacky-ish way of rendering legends outside of
the respective chart block component by injecting a `renderLegend` prop.
This allows us to render the legend in the parent `ChartRenderer` and
avoids the rendered chart being pushed down.

The rendered chart will now actually respect the specified height that
is set in the chart builder.