EES-1399 Created data files to test duplicate location entries
If you upload a data file with the same location code as a previously
existing location but use a different location name, then now the
importer will create a duplicate entry. These two sets of subject data
files allow you to test that entries in the Statistics DB Location table
are duplicated as expected.