BAU-400 use Formik and reducer state for ChartBuilder components
This commit performs significant refactoring of the current
`ChartBuilder` to dramatically simplify its state model.

Currently, it is driven by a series of `useState` and `useEffect` hooks
that create a tangled web of dependencies that make it very difficult
to rationalise what's going on (particularly in `ChartBuilder`).
Additionally, there is no proper form validation of the the constituent
form components so there is a real issue that we can break the
consistency of our chart configuration.

To solve these issue, the constituent form components are now driven by
Formik and some minor validation (potentially more to be added). Changes
are propagated back up to `ChartBuilder` via a new `Effect` component
(as Formik does not this natively).

New chart changes are stored in state via `chartBuilderReducer`. This
primarily allows us to consolidate all of our state changes in one
location (previously, state was being changed in all number of
locations).

From this we can just derive new slices of state using `useMemo`,
drastically simplifying how we propagate state down to our chart
preview.