BAU-400 Unify/improve chart types and simplify chart definitions
This commit unifies chart types into common's `chart.ts` instead of
having them spread between `publicationService.ts` and `chart.ts`.

As part of this we refactor `ChartRendererProps` to use a discriminated
union instead of an intersection between all the different chart props.

This allows the compiler to more strictly enforce checking of the
different chart types so that we more precisely build up the
correct props for that chart. This is useful in `ChartBuilder` for
example.

We also enrich the types for charts like `LineChartBlock` by providing
the exact `axes` that are allowed.

Chart definitions have been refactored to use `defaults` and
`constants` bags. This makes it easier to merge the definition's options
together in `ChartBuilder`,