EES-1303 Ensure data blocks with one result will appear in list of possible key stats
This fixes the bug where data blocks that returned one result didn't
appear in the list of data blocks available to add as a key stat tile.
This bug occurred because in a data block request, you only have a flat
array of filter GUIDs. This doesn't tell you whether those filters
belong to the same category or not, and therefore it is not possible from
the response alone to tell whether that data block will return one or
more results.

My quick (imperfect) fix is continue to filter data blocks that contain more than
one timePeriod or indicator, but allow data blocks with any number of
filters. This means that data blocks that contain multiple results may
still appear as an option for a key stat tile if they have multiple
filters from the same filter group selected.