BAU-426 Refactor DataBlock dependent components into DataBlockTabs
This commit aims to unify all the disparate rendering implementations
for a `DataBlock` into a single component called `DataBlockTabs`.

The idea is that `DataBlockTabs` is flexible enough to handle all use
cases so that we can avoid having essentially the same component being
tweaked slightly for different situations e.g. for data blocks in
release headlines vs the main release content.

Achieving this also has had the effect of refactoring `SectionBlocks`
and removing `BlockRenderer`. `BlockRenderer` is a bit of a nasty
component as it tries to unify render both `ContentBlock` and
`DataBlock` types. This makes it prone to prop bloat and having to make
most props optional.

A better approach was to split out `ContentBlockRender` and use this as
a replacement for sections that only accept `ContentBlocks` anyway.