EES-1565 Hoist control of block ordering out of EditableSectionBlocks
This commit converts `EditableSectionBlocks` into a controlled component
so that block state is maintained in the parent component. This
simplifies the code and also allows us to optimize performance as
re-ordering the blocks was typically causing unnecessary re-renders.

Re-renders are problematic as they can cause `EditableDataBlock` to
re-render, which causes new data block queries to be triggered (can
be quite expensive on the data API).