refactor: re-implement Accordion open state using dictionary
This replaces the current Accordion state implementation for tracking
open sections to a `Dictionary<boolean>`. This is important as we cannot
reliably track which sections are open if the order of the children
changes. This is necessary when using an `EditableAccordion` where order
can indeed change.

This commit also begins to memoize and cleanup some of the general
internals for `Accordion`. As part of this we now allow a `openAll`
prop that can trigger all sections to open/close.