feat: always show editable block buttons using EditableBlockWrapper
This changes allows us to significantly simplify and remove a lot of
the individualy editable block renderer components as we no longer
have to worry about dealing with the buttons within.

Ensuring the buttons are always visible is also an improvement in
accessibility as we don't have to deal with the them only being
visible on hover (in content blocks).