BAU-71 Pass more detailed row/col headers down to MultiHeaderTable
This commit replaces the current header implementation which uses a
simplistic model where `MultiHeaderTable` expands the headers out for
us. This is a limitation as it expects the headers to always expand to
a specific configuration (the full cartesian set).

To make headers more flexible, we now expand out most of the header
details in the parent component `TimePeriodDataTable` and pass these
down. This allows us to perform additional transformations to our
headers that wouldn't have been possible before.

A lot of the code has needed refactoring to make it work with this new
paradigm. Some of it may be more or less complicated, but it's hard to
know given the general ambient complexity with generating these tables.