BAU-468 Fix timing issue causing Details to fail to open when searching
I believe the `Details` could fail to open as React could potentially be
in the middle of re-rendering the component, meaning that the click
event could potentially be lost.

Adding delays simply slows down the code enough that React has time to
fully update the DOM before we try and do some subsequent step(s).