EES-616 Fix incorrect matching of data sets to configuration and categories
This commit re-implements the internal expansion of data sets so that we
only generate extra data sets for incomplete data sets (e.g. where
time period and/or location is missing).

We also need to pass data sets through as a `ChildDataSet` so that we
can retain the original 'parent' `DataSet` that the user configured.
This is important for creating the data set key.

This code is currently quite confusing to think about, so it would be
good to revisit this in the future and figure out if we can simplify it.