EES-616 Only allow default sorting of data sets
This commit comments out any sorting functionality in the UI for the
time being as we don't really know how this should be implemented.

Currently we should just rely on the default algorithm, which sorts the
major axis alphabetically. We then also sort the data sets with in a
category.