EES-729 Perform initial validation of chart forms to ensure correct chart builder `isValid`
Currently, whenever a field changes it changes the chart builder's state
(due to the `Effect` component) and passes through new initial values to
each chart form.

As each Formik instance has `enableReinitialize = true`, this means that
initial values is updated correctly for the form. Unfortunately, the
side effect of this is that `isValid` is always set to false and it
isn't possible to submit the form when the user is on the 'Data' tab of
the chart builder.

To get around this, we can just perform the validation of the initial
values whenever they change (or on first render).