EES-735 Fix moving between updating/creating data blocks throwing error
This commit significantly refactors `ReleaseManageDataBlocksPage` so
that we can extract out a `ReleaseManageDataBlocksPageTabs` component.

This is important as it allows us to separate fetching of the data
blocks from the table data query. This allows us to avoid having to
synchronize loading states as we can just not render
`ReleaseManageDataBlocksPageTabs` until all data blocks have been
rendered.

To fix the main ticket issue, we attach a `key` to
`ReleaseManageDataBlocksPageTabs` using the selected data block (or
undefined if there isn't one). This forces React to completely re-render
the tabs allowing us to avoid stale state issues.