EES-735 Fix final step `response` not updating correctly when table is changed
This fix prevents a stale table data response being passed through to
`TableTool`s `finalStep` render prop. We do this by un-setting the
`response` state before we perform the table query.

As part of this, we've also refactored `DataBlockSourceWizard` to have a
final step component like `TableTool`. This makes it easier to
synchronize table header changes as this state can be maintained by the
respective final step component (instead of the wizard itself).