EES-735 Prevent reloading of Table/Chart tabs when data block is saved
Currently, the Table and Chart tabs have to be completely re-rendered
whenever the data block is saved (e.g. when table headers are changed or
chart is saved). This was a little bit of a suboptimal user experience
so we have improved this by keeping the user on the tab after saving.