EES-807 Sync infographic deletion with ChartBuilder state and onChartSave
Currently, deleting the infographic will not be properly synchronized
with the `ChartBuilder` state meaning that the `fileId` may still be
set when `onChartSave` is called.

We now fix this by using an effect to save the chart. This effect is
called by toggling the ChartBuilder's `shouldSave` state. This allows us
to properly ensure that the `onChartSave` receives the correct
`chartProps`.