EES-801 Replace `withoutErrorHandling` with `unhandledrejection` listener
Currently `withoutErrorHandling` is unreliable as it can create race
conditions where multiple invocations can reset the default error
handling behaviour once the associated task has completed. It's
consequently possible for the user to still see the error page(s) when
using `withoutErrorHandling` in multiple places.

The previous solution was to use `catch`s everywhere, and this can
robustly solve the problem, but is quite verbose.

Our new solution is to attach the default error handling to the
`unhandledrejection` window event instead of the `Client` error handler.
This means that most API requests no longer have a default `catch` that
cannot be easily overridden.

A very nice benefit is that we can now idiomatically add custom error
handling in our code as we no longer need to wrap the request with
something like `withoutErrorHandling`. If no `catch` is added, it falls
through to the `unhandledrejection` event handler which shows the
relevant error page.