EES-854 Data API tests - Remove requests with large responses
This has been done so that timeout failures don't happen on Dev.
These slow responses will be quicker on prod due to better servers,
and performance issues will be noticed when generally testing the
service or by users. So it was decided to remove these tests so that
the automated tests could be more useful rather than constantly be
plagued with false positives.