EES-932 Force server-side renders (reverting Next static optimisations)
This change reverts the static optimisations implemented by EES-892 as
it appears that they cause strange behaviour when it comes to the
injection of environment variables. It appears that loading the app from
a statically generated page (like the home page), will cause the
environment variables not to be loaded in later when client-side
rendering takes place.

This could be potentially remedied by re-loading the environment
variables upon client-side mounting, but for the time being, it seems to
just be safer to force server-side rendering on all pages.