EES-620 Add custom axis labels for charts
This implements the axis labels as elements outside of the chart SVG to
allow us to make use of positional styling that isn't available in SVGs.
By doing this, we've been able to create labels that are much more
responsive and don't have the overlapping issues that are typically
observed when using the default Recharts labels.