EES-620 Extract ChartBuilderPreview to handle re-rendering of chart
This commit extracts the `ChartBuilderPreview` component to enable us to
better handle re-rendering of the chart when chart settings are changed.

Currently, the chart can potentially render with overlapping axis labels
(i.e. when using rotated Y axis labels) as the chart does not completely
re-render. We are able to fix this by assigning a new `key` prop (we're
using a `renderCount` ref).

Unfortunately, this approach leads to a fair bit of jank when the user
is changing chart settings. The forced re-render causes the chart to
momentarily disappear and elements further down the page move up to fill
the space. To the user, this appears like the UI is flickering and isn't
particularly nice.

To get around this, we've set the height of the chart's container
explicitly using its child elements to prevent the displacement of the
rest of the page. There is still a bit of a flicker, but it's now much
less noticeable.