refactor: remove `generateIdList` in favour of specific section ids
We've remove the pre-baked `content-section-{number}` ids generated by
`generateIdList` in favour of ids that are actually specific to the
section they actually involve.

It's a little weird that we implemented the ids in the previous way as
it wasn't particularly convenient or useful. The ids would also not make
any sense in their particular contexts.