refactor: use radio/checkbox values directly in ids without kebabCase
This commit prevents the current `kebabCase` driven implementation from
creating unnecessarily ugly ids when the `value` is a uuid. We are
currently generating ids that look like:
`803-fbf-56-600-f-490-f-8409-6413-a-891720-d`.

One caveat is that we still need to replace whitespace from the id. For
this we have used a simple `replace`.