EES-620 Prevent axis labels from becoming too wide
This change addresses axes with explicitly set widths causing the chart
to break out of its container.

We have also wrapped the x axis label and chart elements with a div so
that they are grouped as one unit. This has the benefit of ensuring the
x axis label remains centered in relation to the chart.