EES-950 Add explicit setting of env variables for clients/services
We're currently seeing some issues that suggest that runtime config
variables on `process.env` aren't being set correctly.
This may be a timing issue with NextJS and how it loads modules on the
client-side render. I suspect the services/clients are being prematurely
initialised before the public runtime config variables are set into the
environment.

To get around this, we're now opting to load in the public runtime
config variables at every render phase. Additionally, we explicitly
initialise the various tracking services on the mount phase instead of
relying on global module-level initialisation.