EES-950 Remove `baseUrls` and `setApiBaseUrls` from api/index.ts
We have replaced `setApiBaseUrls` with explicitly setting individual
client `baseURL`s via `client.axios.defaults.baseURL` as this is a finer
grained approach.

We've also removed `baseUrls` as this is a potential source of problems
when trying to identify if the respective environment variables are
being used in a component or not.