EES-1020 Move checkbox subgrouping into FormCheckboxSearchSubGroups
If there is only one subgroup, we will now render the
`BaseFormCheckboxGroup` in `FormCheckboxSearchSubGroups` instead of
`FormFieldCheckboxSearchSubGroups`. This reduces complexity of the
`FormField` variant considerably and makes `FormCheckboxSearchSubGroups`
more re-usable in non-Formik contexts.

As part of this, we also ensure that there is always a 'Select all
options' button when there is a single subgroup. This avoids cases where
the button is potentially missing.