EES-1148 Fix public frontend not rendering when JS is disabled
This was due to the conditional child rendering that was taking place in
`ApplicationInsightsContextProvider`. We have been doing this primarily
to prevent `useApplicationInsights` from returning an optional
`ApplicationInsights`.

We've now allowed `ApplicationInsights` to be optional and consequently
`ApplicationInsightsContextProvider` will always render its children.
This has meant some tweaks to our code where we `useApplicationInsights`
to avoid potential null pointer errors.