EES-1079 Fix incorrect PascalCasing of filter group keys
This fixes filter group keys not being PascalCased correctly so that
non-alphanumeric characters are filtered out. This is leading to
characters such as dots appearing in the keys.

These keys are used for the Formik field names and unfortunately cause
validation to fail due to the dots being considered as object path
notation by Formik.

We have now corrected the `PascalCase` function (and ergo the
`CamelCase` function) so that only alphanumeric characters are allowed.