EES-1155 Fix null pointer from release headline's with no data block
We've also renamed `DataBlockRenderer` to `DataBlockTabs` as this makes
it clearer that we will be rendering some `Tabs`. This is  motivated by
the release headlines components now render `Tabs` instead of
`DataBlockTabs` if there is no data block to prevent us needing to pass
an optional `dataBlock` prop to `DataBlockTabs`.