EES-974 Stop ChartBuilderPreview height being locked after initial render
This removes code that was originally put in place to try and avoid jank
in the form UI whenever the chart height changed. Unfortunately, it causes
us other issues in that the preview container height gets locked upon
initial render meaning that further changes will not cause the container
to resize correctly. This can make the chart look likes it is breaking out
of the container for example.

We've just removed this code completely as there isn't really that much
jankiness caused by this specific concern (there is other types of jank
though such as chart re-draws).