EES-1229 Add `useStorageItem` hook and sync methods to `storageService`
This adds synchronous methods to `storageService` so that we can
integrate it more easily with `useStorageItem`. These shouldn't be used
if you can avoid it as the async variants will be more performant.

We've also improved the mocking of localStorage for tests with an
in-memory implementation. This mock is reset on every test.