EES-237 Refactor ThemeService GetMyThemes method into GetThemes
The current `GetMyThemes` is needless specialisation for the concept of
'me'. We should only need a single `/themes` endpoint for getting the
list of themes that a user can see.

We also use this opportunity to map `Theme` and `Topic` to their
respective view models, instead of the current partially hydrated
entities (which are misleading).