fix: change PublicationSummary to table to fix 'Releases' row sizing
This fixes the 'Releases' row typically not being wide enough due to the
new actions that have been placed to the right. This is because
`SummaryList` actually uses a `display: table` based implementation
that lacks any sort of colspan type styling.

With a table, we can now correctly size the 'Releases' row to colspan 2.
As table and `display: table` are semantically treated the same by
screen readers, this change should have no actual impact on how the
page is experienced.