EES-1269 Fix chart builder's infographic preview not working
This was being caused by a regression from the recent changes to swap
out usages of infographic filenames for file ids.

We've reverted the relevant changes and instead of trying to update an
existing chart file id, we now just upload the infographic as a new
chart file and then rely on it being replaced when the data block is
updated.