EES-357 Remove chart definition constants in favour of defaults
Currently chart definitions allow you to specify constant values that
will always be set in the current chart/axes options. Whilst this was a
nice idea on paper, this causes issues when trying to swap between
charts.

We've specifically found that changing from/to a map chart causes the
`groupBy` option to be forced to 'locations'. This isn't a major problem
but doesn't make a very nice user experience as the user will most
likely not want this groupBy.

Instead of defining constants in the chart, we can just set these in the
chart component itself. For exmaple, we are already doing this in
`MapBlockInternal` to force the `groupBy` to 'locations'.