EES-1269 Reorganise key stat components
This change tries to reorganise the key stat components so that
`KeyStatColumn` no longer needs to wrap each `KeyStat`. Instead,
`KeyStat` (or `EditableKeyStat`) renders with the `KeyStatColumn` as
one of its children automatically.

For situations where we need more control, we still export
`KeyStatColumn` and `KeyStatTile` e.g. for `MapBlock` indicators.

The advantage of doing this is that it makes it easier to keep the
test ids consistent across key stat components.