EES-1145 Make `Release.PreviousVersionId` field nullable
This changes `Release.PreviousVersionId` to nullable so that we can make
it easier to bulk delete releases when deleting themes/topics. Currently
`PreviousVersionId` cannot be nulled and EF is consequently unable to
delete all of a theme/topic's releases in a single operation due to the
self-referential nature of the `Release` -> `PreviousVersion` relationship.

Additionally, `PreviousVersionId` can take on non-sensical IDs such as
`Guid.Empty` or its own `Id` when it is the first release (doesn't make
sense as the previous version shouldn't be itself). These changes should
correct the semantics of `PreviousVersion` and `PreviousVersionId` i.e.
make them properly optional.