EES-1250 Fix data uploads showing conflicting server validation errors
This change fixes validation errors showing simultaneously for both ZIP
and CSV fields when the user should only see messages relevant to either
depending on their `uploadType` choice.

We've fixed this by allowing the `useFormSubmit` error mappers parameter
to optionally be a function type that receives the form values before
returning the mappings.