EES-1347 Render read-only FormEditor value on IE
This changes `FormEditor` to actually render its HTML contents in a div
styled like a read-only `FormTextArea`. Previously we were not rendering
the contents at all and the field would only show an empty `textarea`.

This conveniently allows us to circumvent needing to pass through a
`name` prop as well.