EES-351 Refactor admin app config to use `ConfigContext`
Currently we rely on a simple `useConfig` hook that tries to load the
app config lazily rather than eagerly. This is slightly beneficial from
a performance standpoint (one less request on initial app render).

Unfortunately, it isn't great from a DX perspective as:
- We need to handle the config potentially being unset
- Need to mock the backend client directly in test code.

This changes the implementation to be driven by a `ConfigContext` which
eagerly loads the config and allows us to create a test
`ConfigContext` more easily.