EES-1367 Refactor most of FileStorageUtils into BlobStorageService
This change removes most dependencies on `FileStorageUtils`
for accessing blob storage. This was clearly problematic as it:

- Exposed too many underlying details of the Azure SDK. This meant that
  an upgrade from v11 to v12 is non-trivial due to breaking changes
  across the entire library.
- Made unit testing impossible as clients are instantiated on a per
  method invocation basis.

We've fixed most of these issues by moving most of the functionality
from `FileStorageUtils` into a new `BlobStorageService`. This class
provides wrapping methods that abstract away the internal client
implementation as much as possible. This should enable us to actually
upgrade the Azure SDK without having to fix breaking changes everywhere.