refactor: rename Either's Unit returning methods to OnSuccessVoid
This change is required as a C# doesn't actually support return type
inferrence (rubbish). There are a bunch of cases where we've been
correctly using `OnSuccess` but it incorrectly infers that we're
using one of the `Either<Tl, Unit>` methods.

We can avoid this problem by calling an `OnSuccessVoid` method instead.