EES-221 Update replacement plan to match frontend requirements
The replacement plan has now been updated to return its fields with
more details, specifically with any labels that can be shown to the
user to make it easier to identify what is missing.

For efficiency, we now also group the various filters and indicators
where possible as well. This is pretty similar to the subject meta
returned from table builder APIs.

The endpoint response is now actually a summary variant of the
replacement plan that removes the details about any valid data blocks
or footnotes. As these are valid, the user doesn't need to know what
combination of filters were selected for example.