EES-221 Add ReleaseDataFilePage to display replacement plan
This change adds the new `ReleaseDataFilePage` that can be navigated to
by clicking on the 'Replace data' link for an uploaded data file.

To view a data replacement plan, you currently have to add a
`replacementFileId` query parameter that corresponds to the file id of
another uploaded file. This will be changed in upcoming work so that the
replacement file can be uploaded directly from this page.