EES-1399 Unset replacement file's `ReplacingId` to allow deletion
Currently the replacement data file's `ReplacingId` field can remain
set if there are multiple releases that are holding onto references to
the original data file (e.g. in amendments).

Unfortunately, this means that when we try to cleanup test releases
and their associated files, we get a foreign key error due to the
`ReplacingId` file reference still existing.