EES-1426 Fix uploaded zip data file returning incorrect details
This fixes the `ReleaseFilesService.UploadDataFilesAsZip` method
returning incorrect `DataFileInfo`. We have now adjusted this to return
details of the zipped data file instead of the ZIP file itself.

As part of this, we have refactored other service APIs involving
`ZipArchiveEntry` parameters to take a `DataArchiveFile` instead. This
allows us to more easily mock things for unit tests.