EES-1429 Allow data block table tool to initialise from invalid queries
This implementation requires us to perform multiple calls to
`tableBuilderService.filterSubjectMeta` for each step of the table tool.

Unfortunately, we serialise the entire data block query as-is from the
user, and we don't filter out any invalid options. This means that
there isn't really a good way for us to assert that a query is actually
valid or not. The only current way is to repeatedly filter the query
upon initialisation as if we were using the table tool normally.