EES-1429 Re-initialise data block table tool with simpler checks
This replaces the previous implementation where we performed multiple
subject meta filter requests to try and accurately pinpoint where the
problem is. Instead, we're now going for a looser approach where we just
simply check that a table can actually be rendered. If it can't, we just
kick the user back to the locations step.

We've gone with this approach as it avoids needing to make multiple
requests upon initialisation. This could really slow down the rendering
of the table tool as some meta queries can take a long time.

A better, more comprehensive solution has been outlined in the ticket,
but we will hold off implementation until we that its worth doing it.