EES-711 Fix inconsistent table header dragging behaviour
This change simply prevents the parent `FormFieldSortableListGroup` from
being dragged whilst the user's mouse is over one of the child
`FormSortableList` rows. This way, there is never any ambiguity over
which element should be dragged, and the user will be able to correctly
drag the table header (instead of the table header group).