EES-352 Merge MetaGuidance update endpoints into single endpoint
This removes the `UpdateSubject` endpoint and merges its behaviour with
the previous `UpdateRelease` endpoint. We now have a single endpoint
`PATCH /release/{releaseId}/meta-guidance` that allows us to do bulk
updates of all subjects.

This change is necessary as the UI is changing towards making only a
single request for all meta guidance updates.