EES-352 Add FormEditor focus/styling improvements
This changes `FormEditor` so that it watches on `hashchange` events so
that it can grab focus if the hash id matches its `id` prop. This
emulates how other form inputs work.

We have also tried to update the label to better emulate how standard
form labels work. This means that clicking the label now focuses the
editor.

Additionally, we have also updates the `FormEditor` styling so that the
focused state is consistent with other form inputs.