EES-352 Fix FormEditor stealing focus upon init
This changes `FormEditor` so that it doesn't steal focus as soon as it
initialises. This was needed originally so that some of the editors e.g.
on the release 'Content' page could be focused when the parent text
block was clicked.

We now enable this behaviour via a `focusOnInit` prop.