EES-1551 Separate pre-release access list endpoints
This change splits out new endpoints for fetching the pre-release access
list. Previously, there would be a large amount of overfetching as we
were fetching the entire `ReleaseViewModel` and the entirety of its
content.

We now just fetch a leaner `PreReleaseAccessListViewModel` that extends
a `ReleaseSummaryViewModel`, which omits most of the content from the
full `ReleaseViewModel`.

As part of this, we have also cleaned and corrected the common
`publicationService` types so that the `Release` type does not have any
frontend/admin specific properties.