EES-1604 Fix UI test teardown throwing foreign key constraint error
This changes the deletion of orphaned subjects so that they are
untracked by EF beforehand. This allows us to avoid EF trying to cascade
the delete to related entities, such as Filters and Observations, in the
wrong order.

Instead, we just execute the deletion of the subject by itself and let
the database-level cascade's handle the rest.