EES-1605 Enforce consistent decimal points for formatted numbers
This change updates `formatPretty` so that its `decimalPoints` parameter
is the exact number of decimal points that the formatted number will
have. This removes any ambiguity of cases where it could still return
a formatted number with fewer decimal places.

As part of this, we have removed 2 decimal places being the default and
replaced it with no decimal places. We also now determine the number of
decimal places to be used in each chart's minor axis based on the data
sets being used.