EES-1605 Re-add 2 d.p as default for map legend groups and indicators
This change reinstates 2 decimal places as the default for
`formatPretty` (and consequently all indicator formatting). This will
have implications for any indicators where analysts have not specified
decimal places e.g. indicators that are actually integers will be
'incorrectly' formatted to 2 decimal places. The advice there is to
ensure that decimal places are explicitly specified in the metadata.