EES-1605 Calculate default decimal places for map legend groups from values
This change adds additional logic to calculate the correct number of
decimal places to use in the map block legend. We look through all the
values to find the decimal places (up to a max of 2, but always at least
0).

Once we have the decimal places, we increase it by 1 if it would be
insufficient to prevent the legend groups from overlapping. This
means that we actually support a max of 2-3 decimal places if none has
been specified by the metadata.