EES-1636 Fix admin release files not specifying correct file name
This bug was causing the causing browser to infer the file type based
on the blob returned from the API. It seems like most files seem to have
been uploaded with Excel content types, so it would incorrectly infer
file types as xls.

By setting the filename correctly, we provide the file with a .csv
extension, which is enough to make sure the file downloads correctly.