EES-1319 Fix legend items not generating correctly for maps
This change also removes the `fixedAxisGroupBy` capability with the
`constants.groupBy` property. Unlike the previous implementation of
`constants`, we will not be merging this into the chart builder's state
when the chart type changes. Instead, it will simply be used as required
to allow the definition to override any current state.