EES-1531 Use release checklist to validate updated release
This change re-works `ReleaseService.UpdateRelease` to re-use
`ReleaseChecklistService` to perform final validations before a release
can be approved/published. Importantly, we have done this to avoid
having divergent validation rules in the future.

As part of this, we have updated related validation messages so that
they are in-line with the checklist error codes. We want to avoid having
discrepancies as this should make it easier to ensure they are all
mapped correctly in the backend/frontend.